kafka-control-center/backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0

confluent-kafka==2.3.0

pydantic==2.6.4
pydantic-settings==2.2.1

PyYAML==6.0.1

python-dotenv==1.0.1